update rubikscube model

+++ puzzle/world_model/util.py
@@ -21,4 +21,4 @@ def download_dataset():
 def round_through_gradient(x: chex.Array) -> chex.Array:
     # x is a sigmoided value in the range [0, 1]. Use a straight-through estimator:
     # the forward pass returns jnp.round(x) while the gradient flows as if it were the identity.
-    return x + jax.lax.stop_gradient(jnp.round(x) - x)
+    return x + jax.lax.stop_gradient(jnp.round(x).astype(jnp.float32) - x)
